remove old event screen
