+++ scripts/screens/EventsScreen.py
@@ -1,25 +1,20 @@
 import pygame
 import pygame_gui
 
+from scripts.cat.cats import Cat
 from scripts.event_class import Single_Event
 from scripts.events import events_class
-from scripts.game_structure.game_essentials import game, screen_x, screen_y, MANAGER
-from scripts.game_structure.ui_elements import IDImageButton, UIImageButton, UIModifiedScrollingContainer
+from scripts.game_structure import image_cache
+from scripts.game_structure.game_essentials import game, MANAGER
+from scripts.game_structure.ui_elements import UIImageButton, UIModifiedScrollingContainer, IDImageButton
 from scripts.game_structure.windows import GameOver
-from scripts.utility import (
-    get_living_clan_cat_count,
-    get_text_box_theme,
-    scale,
-    shorten_text_to_fit,
-    clan_symbol_sprite,
-)
-from .Screens import Screens
-from ..cat.cats import Cat
-from ..game_structure import image_cache
+from scripts.screens.Screens import Screens
+from scripts.utility import scale, clan_symbol_sprite, get_text_box_theme, shorten_text_to_fit, \
+    get_living_clan_cat_count
 
 
 class EventsScreen(Screens):
-    event_display_type = "all events"
+    event_display_type = "all"
     all_events = ""
     ceremony_events = ""
     birth_death_events = ""
@@ -30,36 +25,26 @@ class EventsScreen(Screens):
     display_text = (
         "<center>See which events are currently happening in the Clan.</center>"
     )
-    display_events = ""
+    display_events = []
 
-    def __init__(self, name=None):
+    def __init__(self, name):
         super().__init__(name)
-        self.clan_symbol = None
-        self.misc_alert = None
-        self.other_clans_alert = None
-        self.health_alert = None
-        self.relation_alert = None
-        self.birth_death_alert = None
-        self.ceremony_alert = None
-        self.misc_events_button = None
-        self.other_clans_events_button = None
-        self.health_events_button = None
-        self.birth_death_events_button = None
-        self.ceremonies_events_button = None
-        self.all_events_button = None
-        self.relationship_events_button = None
-        self.events_list_box = None
-        self.toggle_borders_button = None
-        self.timeskip_button = None
-        self.events_frame = None
-        self.clan_age = None
-        self.season = None
-        self.heading = None
-        self.display_events_elements = {}
-        self.involved_cat_buttons = []
-        self.cat_profile_buttons = {}
-        self.scroll_height = {}
+
         self.events_thread = None
+        self.event_screen_container = None
+        self.clan_info = {}
+        self.timeskip_button = None
+
+        self.full_event_display_container = None
+        self.events_frame = None
+        self.event_buttons = {}
+        self.alert = {}
+
+        self.event_display = None
+        self.event_display_elements = {}
+        self.cat_profile_buttons = {}
+        self.involved_cat_container = None
+        self.involved_cat_buttons = {}
 
         # Stores the involved cat button that currently has its cat profile buttons open
         self.open_involved_cat_button = None
@@ -68,244 +53,89 @@ class EventsScreen(Screens):
 
     def handle_event(self, event):
         if game.switches["window_open"]:
-            pass
-        elif event.type == pygame_gui.UI_BUTTON_ON_HOVERED:
-            if (
-                    event.ui_element == self.ceremonies_events_button
-                    and self.ceremony_alert
-            ):
-                self.ceremony_alert.kill()
-            elif (
-                    event.ui_element == self.birth_death_events_button
-                    and self.birth_death_alert
-            ):
-                self.birth_death_alert.kill()
-            elif (
-                    event.ui_element == self.relationship_events_button
-                    and self.relation_alert
-            ):
-                self.relation_alert.kill()
-            elif event.ui_element == self.health_events_button and self.health_alert:
-                self.health_alert.kill()
-            elif (
-                    event.ui_element == self.other_clans_events_button
-                    and self.other_clans_alert
-            ):
-                self.other_clans_alert.kill()
-            elif event.ui_element == self.misc_events_button and self.misc_alert:
-                self.misc_alert.kill()
-
-        if game.switches["window_open"]:
             return
 
-        if event.type == pygame_gui.UI_BUTTON_START_PRESS:
-            if event.ui_element == self.timeskip_button:
+        if event.type == pygame_gui.UI_BUTTON_ON_HOVERED:
+            element = event.ui_element
+            if element in self.event_buttons.values():
+                for ele in self.event_buttons:
+                    if ele == "all":
+                        continue
+                    if self.event_buttons[ele] == element:
+                        x_pos = int(self.alert[ele].get_relative_rect()[0] - 10)
+                        y_pos = self.alert[ele].get_relative_rect()[1]
+                        self.alert[ele].set_relative_position((x_pos, y_pos))
+
+        if event.type == pygame_gui.UI_BUTTON_ON_UNHOVERED:
+            element = event.ui_element
+            if element in self.event_buttons.values():
+                for ele in self.event_buttons:
+                    if ele == "all":
+                        continue
+                    if self.event_buttons[ele] == element:
+                        x_pos = int(self.alert[ele].get_relative_rect()[0] + 10)
+                        y_pos = self.alert[ele].get_relative_rect()[1]
+                        self.alert[ele].set_relative_position((x_pos, y_pos))
+
+        if event.type == pygame_gui.UI_BUTTON_START_PRESS:  # this happens on start press to prevent alert movement
+            element = event.ui_element
+            if element in self.event_buttons.values():
+                for ele, val in self.event_buttons.items():
+                    if val == element:
+                        self.handle_tab_event(ele)
+                        break
+
+        if event.type == pygame_gui.UI_BUTTON_PRESSED:  # everything else on button press to prevent blinking
+            element = event.ui_element
+            if element == self.timeskip_button:
                 self.events_thread = self.loading_screen_start_work(
                     events_class.one_moon
                 )
-
-            # Change the type of events displayed
-            elif event.ui_element == self.all_events_button:
-                if self.event_container.vert_scroll_bar:
-                    self.scroll_height[self.event_display_type] = (
-                            self.event_container.vert_scroll_bar.scroll_position
-                            / self.event_container.vert_scroll_bar.scrollable_height
-                    )
-                self.event_display_type = "all events"
-                # Update Display
-                self.update_list_buttons(self.all_events_button)
-                self.display_events = self.all_events
-                self.update_events_display()
-            elif event.ui_element == self.ceremonies_events_button:
-                if self.event_container.vert_scroll_bar:
-                    self.scroll_height[self.event_display_type] = (
-                            self.event_container.vert_scroll_bar.scroll_position
-                            / self.event_container.vert_scroll_bar.scrollable_height
-                    )
-                self.event_display_type = "ceremony events"
-                self.ceremonies_events_button.disable()
-                # Update Display
-                self.update_list_buttons(
-                    self.ceremonies_events_button, self.ceremony_alert
-                )
-                self.display_events = self.ceremony_events
-                self.update_events_display()
-            elif event.ui_element == self.birth_death_events_button:
-                if self.event_container.vert_scroll_bar:
-                    self.scroll_height[self.event_display_type] = (
-                            self.event_container.vert_scroll_bar.scroll_position
-                            / self.event_container.vert_scroll_bar.scrollable_height
-                    )
-                self.event_display_type = "birth death events"
-                self.birth_death_events_button.enable()
-                # Update Display
-                self.update_list_buttons(
-                    self.birth_death_events_button, self.birth_death_alert
-                )
-                self.display_events = self.birth_death_events
-                self.update_events_display()
-            elif event.ui_element == self.relationship_events_button:
-                if self.event_container.vert_scroll_bar:
-                    self.scroll_height[self.event_display_type] = (
-                            self.event_container.vert_scroll_bar.scroll_position
-                            / self.event_container.vert_scroll_bar.scrollable_height
-                    )
-                self.event_display_type = "relationship events"
-                self.relationship_events_button.enable()
-                # Update Display
-                self.update_list_buttons(
-                    self.relationship_events_button, self.relation_alert
-                )
-                self.display_events = self.relation_events
-                self.update_events_display()
-            elif event.ui_element == self.health_events_button:
-                if self.event_container.vert_scroll_bar:
-                    self.scroll_height[self.event_display_type] = (
-                            self.event_container.vert_scroll_bar.scroll_position
-                            / self.event_container.vert_scroll_bar.scrollable_height
-                    )
-                self.event_display_type = "health events"
-                self.health_events_button.disable()
-                # Update Display
-                self.update_list_buttons(self.health_events_button, self.health_alert)
-                self.display_events = self.health_events
-                self.update_events_display()
-            elif event.ui_element == self.other_clans_events_button:
-                if self.event_container.vert_scroll_bar:
-                    self.scroll_height[self.event_display_type] = (
-                            self.event_container.vert_scroll_bar.scroll_position
-                            / self.event_container.vert_scroll_bar.scrollable_height
-                    )
-                self.event_display_type = "other clans events"
-                self.other_clans_events_button.disable()
-                # Update Display
-                self.update_list_buttons(
-                    self.other_clans_events_button, self.other_clans_alert
-                )
-                self.display_events = self.other_clans_events
-                self.update_events_display()
-            elif event.ui_element == self.misc_events_button:
-                if self.event_container.vert_scroll_bar:
-                    self.scroll_height[self.event_display_type] = (
-                            self.event_container.vert_scroll_bar.scroll_position
-                            / self.event_container.vert_scroll_bar.scrollable_height
-                    )
-                self.event_display_type = "misc events"
-                self.misc_events_button.disable()
-                # Update Display
-                self.update_list_buttons(self.misc_events_button, self.misc_alert)
-                self.display_events = self.misc_events
-                self.update_events_display()
-            elif event.ui_element in self.involved_cat_buttons:
-                self.make_cat_buttons(event.ui_element)
-            elif event.ui_element in self.cat_profile_buttons:
-                game.switches["cat"] = event.ui_element.ids
+            elif element in self.involved_cat_buttons.values():
+                self.make_cat_buttons(element)
+            elif element in self.cat_profile_buttons.values():
+                self.save_scroll_position()
+                game.switches["cat"] = element.ids
                 self.change_screen("profile screen")
             else:
+                self.save_scroll_position()
                 self.menu_button_pressed(event)
 
-        elif event.type == pygame.KEYDOWN and game.settings["keybinds"]:
-            if event.key == pygame.K_RIGHT:
-                self.change_screen("camp screen")
-            elif event.key == pygame.K_UP:
-                if self.event_display_type == "ceremony events":
-                    self.event_display_type = "all events"
-                    # Update Display
-                    self.update_list_buttons(self.all_events_button)
-                    self.display_events = self.all_events
-                    self.update_events_display()
-                elif self.event_display_type == "birth death events":
-                    self.event_display_type = "ceremony events"
-                    # Update Display
-                    self.update_list_buttons(
-                        self.ceremonies_events_button, self.ceremony_alert
-                    )
-                    self.display_events = self.ceremony_events
-                    self.update_events_display()
-                elif self.event_display_type == "relationship events":
-                    self.event_display_type = "birth death events"
-                    # Update Display
-                    self.update_list_buttons(
-                        self.birth_death_events_button, self.birth_death_alert
-                    )
-                    self.display_events = self.birth_death_events
-                    self.update_events_display()
-                elif self.event_display_type == "health events":
-                    self.event_display_type = "relationship events"
-                    # Update Display
-                    self.update_list_buttons(
-                        self.relationship_events_button, self.relation_alert
-                    )
-                    self.display_events = self.relation_events
-                    self.update_events_display()
-                elif self.event_display_type == "other clans events":
-                    self.event_display_type = "health events"
-                    # Update Display
-                    self.update_list_buttons(
-                        self.health_events_button, self.health_alert
-                    )
-                    self.display_events = self.health_events
-                    self.update_events_display()
-                elif self.event_display_type == "misc events":
-                    self.event_display_type = "other clans events"
-                    # Update Display
-                    self.update_list_buttons(
-                        self.other_clans_events_button, self.other_clans_alert
-                    )
-                    self.display_events = self.other_clans_events
-                    self.update_events_display()
-            elif event.key == pygame.K_DOWN:
-                if self.event_display_type == "all events":
-                    self.event_display_type = "ceremony events"
-                    # Update Display
-                    self.update_list_buttons(
-                        self.ceremonies_events_button, self.ceremony_alert
-                    )
-                    self.display_events = self.ceremony_events
-                    self.update_events_display()
-                elif self.event_display_type == "ceremony events":
-                    self.event_display_type = "birth death events"
-                    # Update Display
-                    self.update_list_buttons(
-                        self.birth_death_events_button, self.birth_death_alert
-                    )
-                    self.display_events = self.birth_death_events
-                    self.update_events_display()
-                elif self.event_display_type == "birth death events":
-                    self.event_display_type = "relationship events"
-                    # Update Display
-                    self.update_list_buttons(
-                        self.relationship_events_button, self.relation_alert
-                    )
-                    self.display_events = self.relation_events
-                    self.update_events_display()
-                elif self.event_display_type == "relationship events":
-                    self.event_display_type = "health events"
-                    # Update Display
-                    self.update_list_buttons(
-                        self.health_events_button, self.health_alert
-                    )
-                    self.display_events = self.health_events
-                    self.update_events_display()
-                elif self.event_display_type == "health events":
-                    self.event_display_type = "other clans events"
-                    # Update Display
-                    self.update_list_buttons(
-                        self.other_clans_events_button, self.other_clans_alert
-                    )
-                    self.display_events = self.other_clans_events
-                    self.update_events_display()
-                elif self.event_display_type == "other clans events":
-                    self.event_display_type = "misc events"
-                    # Update Display
-                    self.update_list_buttons(self.misc_events_button, self.misc_alert)
-                    self.display_events = self.misc_events
-                    self.update_events_display()
-            elif event.key == pygame.K_SPACE:
+    def save_scroll_position(self):
+        """
+        adds current event display vert scroll bar position to game.switches["saved_scroll_positions"] dict
+        """
+        if self.event_display.vert_scroll_bar:
+            game.switches["saved_scroll_positions"][self.event_display_type] = (
+                    self.event_display.vert_scroll_bar.scroll_position
+                    / self.event_display.vert_scroll_bar.scrollable_height
+            )
 
-                self.events_thread = self.loading_screen_start_work(
-                    events_class.one_moon
-                )
+    def handle_tab_event(self, display_type):
+        self.save_scroll_position()
+
+        self.event_display_type = display_type
+        self.update_list_buttons()
+
+        if display_type == "all":
+            self.display_events = self.all_events
+        elif display_type == "ceremony":
+            self.display_events = self.ceremony_events
+        elif display_type == "birth_death":
+            self.display_events = self.birth_death_events
+        elif display_type == "relationship":
+            self.display_events = self.relation_events
+        elif display_type == "health":
+            self.display_events = self.health_events
+        elif display_type == "other_clans":
+            self.display_events = self.other_clans_events
+        elif display_type == "misc":
+            self.display_events = self.misc_events
+
+        if display_type != "all":
+            self.alert[display_type].hide()
+
+        self.update_events_display()
 
     def screen_switches(self):
         # On first open, update display events list
@@ -314,513 +144,190 @@ class EventsScreen(Screens):
             self.update_display_events_lists()
             self.display_events = self.all_events
 
-        self.clan_symbol = pygame_gui.elements.UIImage(
+        self.event_screen_container = pygame_gui.core.UIContainer(
+            scale(pygame.Rect((0, 0), (1600, 1400))),
+            object_id="#event_screen_container",
+            starting_height=1,
+            manager=MANAGER
+        )
+
+        self.clan_info["symbol"] = pygame_gui.elements.UIImage(
             scale(pygame.Rect((455, 210), (200, 200))),
             pygame.transform.scale(clan_symbol_sprite(game.clan), (200, 200)),
             object_id=f"clan_symbol",
             starting_height=1,
+            container=self.event_screen_container,
             manager=MANAGER,
         )
 
-        self.heading = pygame_gui.elements.UITextBox(
+        self.clan_info["heading"] = pygame_gui.elements.UITextBox(
             "Timeskip to progress your Clan's life.",
             scale(pygame.Rect((680, 310), (500, -1))),
             object_id=get_text_box_theme("#text_box_30_horizleft_spacing_95"),
+            starting_height=1,
+            container=self.event_screen_container,
             manager=MANAGER,
         )
 
-        self.season = pygame_gui.elements.UITextBox(
+        self.clan_info["season"] = pygame_gui.elements.UITextBox(
             f"Current season: {game.clan.current_season}",
             scale(pygame.Rect((680, 205), (1200, 80))),
-            object_id=get_text_box_theme("#text_box_30_horizleft_spacing_95"),
+            object_id=get_text_box_theme("#text_box_30"),
+            starting_height=1,
+            container=self.event_screen_container,
             manager=MANAGER,
         )
-        self.clan_age = pygame_gui.elements.UITextBox(
+        self.clan_info["age"] = pygame_gui.elements.UITextBox(
             "",
             scale(pygame.Rect((680, 245), (1200, 80))),
-            object_id=get_text_box_theme("#text_box_30_horizleft_spacing_95"),
+            object_id=get_text_box_theme("#text_box_30"),
+            starting_height=1,
+            container=self.event_screen_container,
             manager=MANAGER,
         )
 
-        self.events_frame = pygame_gui.elements.UIImage(
-            scale(pygame.Rect((412, 532), (1068, 740))),
-            image_cache.load_image(
-                "resources/images/event_page_frame.png"
-            ).convert_alpha(),
-            manager=MANAGER,
-        )
-        self.events_frame.disable()
         # Set text for Clan age
         if game.clan.age == 1:
-            self.clan_age.set_text(f"Clan age: {game.clan.age} moon")
+            self.clan_info["age"].set_text(f"Clan age: {game.clan.age} moon")
         if game.clan.age != 1:
-            self.clan_age.set_text(f"Clan age: {game.clan.age} moons")
+            self.clan_info["age"].set_text(f"Clan age: {game.clan.age} moons")
 
         self.timeskip_button = UIImageButton(
             scale(pygame.Rect((620, 436), (360, 60))),
             "",
             object_id="#timeskip_button",
+            starting_height=1,
+            container=self.event_screen_container,
             manager=MANAGER,
         )
 
-        # Sets up the buttons to switch between the event types.
-        self.all_events_button = UIImageButton(
-            scale(pygame.Rect((120, 570), (300, 60))),
-            "",
-            object_id="#all_events_button",
-            manager=MANAGER,
+        self.full_event_display_container = pygame_gui.core.UIContainer(
+            scale(pygame.Rect((90, 532), (1400, 1400))),
+            object_id="#event_display_container",
+            starting_height=1,
+            container=self.event_screen_container,
+            manager=MANAGER
         )
-        self.ceremonies_events_button = UIImageButton(
-            scale(pygame.Rect((120, 672), (300, 60))),
-            "",
-            object_id="#ceremony_events_button",
-            manager=MANAGER,
-        )
-        self.birth_death_events_button = UIImageButton(
-            scale(pygame.Rect((120, 772), (300, 60))),
-            "",
-            object_id="#birth_death_events_button",
-            manager=MANAGER,
-        )
-        self.relationship_events_button = UIImageButton(
-            scale(pygame.Rect((120, 872), (300, 60))),
-            "",
-            object_id="#relationship_events_button",
-        )
-        self.health_events_button = UIImageButton(
-            scale(pygame.Rect((120, 972), (300, 60))),
-            "",
-            object_id="#health_events_button",
-            manager=MANAGER,
-        )
-        self.other_clans_events_button = UIImageButton(
-            scale(pygame.Rect((120, 1072), (300, 60))),
-            "",
-            object_id="#other_clans_events_button",
-            manager=MANAGER,
-        )
-        self.misc_events_button = UIImageButton(
-            scale(pygame.Rect((120, 1172), (300, 60))),
-            "",
-            object_id="#misc_events_button",
+        self.events_frame = pygame_gui.elements.UIImage(
+            scale(pygame.Rect((322, 0), (1068, 740))),
+            image_cache.load_image(
+                "resources/images/event_page_frame.png"
+            ).convert_alpha(),
+            object_id="#events_frame",
+            starting_height=2,
+            container=self.full_event_display_container,
             manager=MANAGER,
         )
 
-        if self.event_display_type == "all events":
-            self.all_events_button.disable()
-        elif self.event_display_type == "ceremony events":
-            self.ceremonies_events_button.disable()
-        elif self.event_display_type == "birth death events":
-            self.birth_death_events_button.disable()
-        elif self.event_display_type == "relationship events":
-            self.relationship_events_button.disable()
-        elif self.event_display_type == "health events":
-            self.health_events_button.disable()
-        elif self.event_display_type == "other clans events":
-            self.other_clans_events_button.disable()
-        elif self.event_display_type == "misc events":
-            self.misc_events_button.disable()
+        y_pos = 0
+        for event_type in ["all", "ceremony", "birth_death", "relationship", "health", "other_clans", "misc"]:
+            self.event_buttons[f"{event_type}"] = UIImageButton(
+                scale(pygame.Rect((30, 38 + y_pos), (300, 60))),
+                "",
+                object_id=f"#{event_type}_events_button",
+                starting_height=1,
+                container=self.full_event_display_container,
+                manager=MANAGER
+            )
 
-        self.misc_alert = None
-        self.other_clans_alert = None
-        self.health_alert = None
-        self.relation_alert = None
-        self.birth_death_alert = None
-        self.ceremony_alert = None
+            if event_type != "all":
+                self.alert[f"{event_type}"] = pygame_gui.elements.UIImage(
+                    scale(pygame.Rect((20, 48 + y_pos), (8, 44))),
+                    pygame.transform.scale(
+                        image_cache.load_image("resources/images/alert_mark.png"), (8, 44)
+                    ),
+                    container=self.full_event_display_container,
+                    object_id=f"alert_mark_{event_type}",
+                    manager=MANAGER,
+                    visible=False
+                )
 
+            y_pos += 100
+
+        self.event_buttons[self.event_display_type].disable()
+
+        self.make_event_scrolling_container()
         self.open_involved_cat_button = None
-        self.make_events_container()
-        self.events_container_y = self.event_container.get_relative_rect()[3]
-
-        # Display text
-        # self.explain_text = pygame_gui.elements.UITextBox(self.display_text, scale(pygame.Rect((25,110),(750,40))))
+        self.update_events_display()
 
         # Draw and disable the correct menu buttons.
         self.set_disabled_menu_buttons(["events_screen"])
         self.update_heading_text(f"{game.clan.name}Clan")
         self.show_menu_buttons()
-        self.update_events_display()
 
-    def exit_screen(self):
-        self.open_involved_cat_button = None
-        self.clan_symbol.kill()
-        self.timeskip_button.kill()
-        del self.timeskip_button
-        self.all_events_button.kill()
-        del self.all_events_button
-        self.ceremonies_events_button.kill()
-        del self.ceremonies_events_button
-        if self.ceremony_alert:
-            self.ceremony_alert.kill()
-            del self.ceremony_alert
-        self.birth_death_events_button.kill()
-        del self.birth_death_events_button
-        if self.birth_death_alert:
-            self.birth_death_alert.kill()
-            del self.birth_death_alert
-        self.relationship_events_button.kill()
-        del self.relationship_events_button
-        if self.relation_alert:
-            self.relation_alert.kill()
-            del self.relation_alert
-        self.health_events_button.kill()
-        del self.health_events_button
-        if self.health_alert:
-            self.health_alert.kill()
-            del self.health_alert
-        self.other_clans_events_button.kill()
-        del self.other_clans_events_button
-        if self.other_clans_alert:
-            self.other_clans_alert.kill()
-            del self.other_clans_alert
-        self.misc_events_button.kill()
-        del self.misc_events_button
-        if self.misc_alert:
-            self.misc_alert.kill()
-            del self.misc_alert
-        self.events_frame.kill()
-        del self.events_frame
-        self.heading.kill()
-        del self.heading
-        self.clan_age.kill()
-        del self.clan_age
-        self.season.kill()
-        del self.season
-        self.event_container.kill()
+    def make_event_scrolling_container(self):
+        if self.event_display:
+            self.event_display.kill()
 
-        for ele in self.display_events_elements:
-            self.display_events_elements[ele].kill()
-        self.display_events_elements = {}
-
-        for ele in self.involved_cat_buttons:
-            ele.kill()
-        self.involved_cat_buttons = []
-
-        for ele in self.cat_profile_buttons:
-            ele.kill()
-        self.cat_profile_buttons = []
-
-        self.hide_menu_buttons()
-
-    def on_use(self):
-
-        self.loading_screen_on_use(self.events_thread, self.timeskip_done)
-
-    def timeskip_done(self):
-        """Various sorting and other tasks that must be done with the timeskip is over."""
-
-        self.scroll_height = {}
-
-        if get_living_clan_cat_count(Cat) == 0:
-            GameOver("events screen")
-
-        self.event_display_type = "all events"
-        self.all_events_button.disable()
-        self.all_events = [
-            x for x in game.cur_events_list if "interaction" not in x.types
-        ]
-
-        self.ceremonies_events_button.enable()
-        if self.ceremony_alert:
-            self.ceremony_alert.kill()
-        self.ceremony_events = [
-            x for x in game.cur_events_list if "ceremony" in x.types
-        ]
-        if self.ceremony_events:
-            self.ceremony_alert = pygame_gui.elements.UIImage(
-                scale(pygame.Rect((110, 680), (8, 44))),
-                pygame.transform.scale(
-                    image_cache.load_image("resources/images/alert_mark.png"), (8, 44)
-                ),
-                manager=MANAGER,
-            )
-
-        if self.birth_death_alert:
-            self.birth_death_alert.kill()
-        self.birth_death_events_button.enable()
-        self.birth_death_events = [
-            x for x in game.cur_events_list if "birth_death" in x.types
-        ]
-        if self.birth_death_events:
-            self.birth_death_alert = pygame_gui.elements.UIImage(
-                scale(pygame.Rect((110, 780), (8, 44))),
-                pygame.transform.scale(
-                    image_cache.load_image("resources/images/alert_mark.png"), (8, 44)
-                ),
-                manager=MANAGER,
-            )
-
-        if self.relation_alert:
-            self.relation_alert.kill()
-        self.relationship_events_button.enable()
-        self.relation_events = [
-            x for x in game.cur_events_list if "relation" in x.types
-        ]
-        if self.relation_events:
-            self.relation_alert = pygame_gui.elements.UIImage(
-                scale(pygame.Rect((110, 880), (8, 44))),
-                pygame.transform.scale(
-                    image_cache.load_image("resources/images/alert_mark.png"), (8, 44)
-                ),
-                manager=MANAGER,
-            )
-
-        if self.health_alert:
-            self.health_alert.kill()
-        self.health_events_button.enable()
-        self.health_events = [x for x in game.cur_events_list if "health" in x.types]
-        if self.health_events:
-            self.health_alert = pygame_gui.elements.UIImage(
-                scale(pygame.Rect((110, 980), (8, 44))),
-                pygame.transform.scale(
-                    image_cache.load_image("resources/images/alert_mark.png"), (8, 44)
-                ),
-                manager=MANAGER,
-            )
-
-        if self.other_clans_alert:
-            self.other_clans_alert.kill()
-        self.other_clans_events_button.enable()
-        self.other_clans_events = [
-            x for x in game.cur_events_list if "other_clans" in x.types
-        ]
-        if self.other_clans_events:
-            self.other_clans_alert = pygame_gui.elements.UIImage(
-                scale(pygame.Rect((110, 1080), (8, 44))),
-                pygame.transform.scale(
-                    image_cache.load_image("resources/images/alert_mark.png"), (8, 44)
-                ),
-                manager=MANAGER,
-            )
-
-        if self.misc_alert:
-            self.misc_alert.kill()
-        self.misc_events_button.enable()
-        self.misc_events = [x for x in game.cur_events_list if "misc" in x.types]
-        if self.misc_events:
-            self.misc_alert = pygame_gui.elements.UIImage(
-                scale(pygame.Rect((110, 1180), (8, 44))),
-                pygame.transform.scale(
-                    image_cache.load_image("resources/images/alert_mark.png"), (8, 44)
-                ),
-                manager=MANAGER,
-            )
-
-        if self.event_display_type == "all events":
-            # if events list is empty, add a single message the says nothing interesting happened
-            if not self.all_events:
-                self.all_events.append(
-                    Single_Event("Nothing interesting happened this moon.")
-                )
-            self.display_events = self.all_events
-        elif self.event_display_type == "ceremony events":
-            self.display_events = self.ceremony_events
-        elif self.event_display_type == "birth death events":
-            self.display_events = self.birth_death_events
-        elif self.event_display_type == "relationship events":
-            self.display_events = self.relation_events
-        elif self.event_display_type == "health events":
-            self.display_events = self.health_events
-        elif self.event_display_type == "other clans events":
-            self.display_events = self.other_clans_events
-        elif self.event_display_type == "misc events":
-            self.display_events = self.misc_events
-
-        self.update_events_display()
-        self.show_menu_buttons()
-
-    def update_list_buttons(self, current_list, current_alert=None):
-        """handles the disabling and enabling of the list buttons"""
-
-        # enable all the buttons
-        self.all_events_button.enable()
-        self.ceremonies_events_button.enable()
-        self.birth_death_events_button.enable()
-        self.relationship_events_button.enable()
-        self.health_events_button.enable()
-        self.other_clans_events_button.enable()
-        self.misc_events_button.enable()
-
-        # disable the current button
-        current_list.disable()
-        if current_alert:
-            current_alert.kill()
-
-    def update_events_display(self):
-
-        self.season.set_text(f"Current season: {game.clan.current_season}")
-        if game.clan.age == 1:
-            self.clan_age.set_text(f"Clan age: {game.clan.age} moon")
-        if game.clan.age != 1:
-            self.clan_age.set_text(f"Clan age: {game.clan.age} moons")
-
-        for ele in self.display_events_elements:
-            self.display_events_elements[ele].kill()
-        self.display_events_elements = {}
-
-        for ele in self.involved_cat_buttons:
-            ele.kill()
-        self.involved_cat_buttons = []
-
-        for ele in self.cat_profile_buttons:
-            ele.kill()
-        self.cat_profile_buttons = []
-
-        # In order to set-set the scroll-bar position, we have to remake the scrolling container
-        self.event_container.kill()
-        self.make_events_container()
-
-        # Stop if Clan is new, so that events from previously loaded Clan don't show up
-        if game.clan.age == 0:
-            return
-
-        # Make display, with buttons and all that.
-        box_length = self.event_container.get_relative_rect()[2]
-        i = 0
-        y = 0
-        padding = 70 / 1400 * screen_y
-        button_size = 68 / 1600 * screen_x
-        button_padding = 80 / 1400 * screen_x
-        for ev in self.display_events:
-            if isinstance(ev.text, str):  # Check to make sure text is a string.
-                self.display_events_elements["event" + str(i)] = (
-                    pygame_gui.elements.UITextBox(
-                        ev.text,
-                        pygame.Rect((0, y), (box_length - 20, -1)),
-                        object_id=get_text_box_theme("#text_box_30_horizleft"),
-                        container=self.event_container.scrollable_container,
-                        starting_height=2,
-                        manager=MANAGER,
-                    )
-                )
-                self.display_events_elements["event" + str(i)].disable()
-                # Find the next y-height by finding the height of the text box, and adding 35 for the cats button
-
-                if i % 2 == 0:
-                    if game.settings["dark mode"]:
-                        self.display_events_elements["shading" + str(i)] = (
-                            pygame_gui.elements.UIImage(
-                                pygame.Rect(
-                                    (0, y),
-                                    (box_length + 100,
-                                     self.display_events_elements["event" + str(i)].get_relative_rect()[3] + padding),
-                                ),
-                                image_cache.load_image("resources/images/shading_dark.png"),
-                                starting_height=1,
-                                container=self.event_container.scrollable_container,
-                                manager=MANAGER,
-                            )
-                        )
-                    else:
-                        self.display_events_elements["shading" + str(i)] = (
-                            pygame_gui.elements.UIImage(
-                                pygame.Rect(
-                                    (0, y),
-                                    (box_length + 100,
-                                     self.display_events_elements["event" + str(i)].get_relative_rect()[3] + padding),
-                                ),
-                                image_cache.load_image("resources/images/shading.png"),
-                                starting_height=1,
-                                container=self.event_container.scrollable_container,
-                                manager=MANAGER,
-                            )
-                        )
-
-                    self.display_events_elements["shading" + str(i)].disable()
-
-                y += self.display_events_elements["event" + str(i)].get_relative_rect()[3]
-
-                self.involved_cat_buttons.append(
-                    IDImageButton(
-                        pygame.Rect(
-                            (self.event_container.get_relative_rect()[2] - button_padding, y - 10),
-                            (button_size, button_size),
-                        ),
-                        ids=ev.cats_involved,
-                        container=self.event_container.scrollable_container,
-                        layer_starting_height=3,
-                        object_id="#events_cat_button",
-                        manager=MANAGER,
-                    )
-                )
-
-                y += 68 / 1600 * screen_y
-                i += 1
-            else:
-                print("Incorrectly formatted event:", ev.text, type(ev))
-
-        # Set scrolling container length
-        # This is a hack-y solution, but it was the easiest way to have the shading go all the way across the box
-        self.event_container.set_scrollable_area_dimensions((box_length, y + 15))
-
-        if self.event_container.vert_scroll_bar:
-            for ele in self.involved_cat_buttons:
-                ele.set_relative_position(
-                    (ele.get_relative_rect()[0] - 20, ele.get_relative_rect()[1])
-                )
-
-        self.event_container.set_dimensions((box_length, self.events_container_y))
-        # Set the scroll bar to the last position it was at
-        if self.scroll_height.get(self.event_display_type):
-            self.event_container.vert_scroll_bar.set_scroll_from_start_percentage(
-                self.scroll_height[self.event_display_type]
-            )
+        self.event_display = UIModifiedScrollingContainer(
+            scale(pygame.Rect((432, 552), (1080, 700))),
+            object_id="#event_display",
+            starting_height=3,
+            manager=MANAGER,
+            allow_scroll_y=True
+        )
 
     def make_cat_buttons(self, button_pressed):
         """Makes the buttons that take you to the profile."""
 
         # Check if the button you pressed doesn't have it cat profile buttons currently displayed.
-        # If it doesn't have it's buttons displayed, set the current open involved_cat_button to the pressed button,
-        # clear all other buttons, and open the cat profile buttons.
-        if self.open_involved_cat_button != button_pressed:
-            self.open_involved_cat_button = button_pressed
-            for ele in self.cat_profile_buttons:
-                ele.kill()
-            self.cat_profile_buttons = []
 
-            pressed_button_pos = (
-                button_pressed.get_relative_rect()[0],
-                button_pressed.get_relative_rect()[1],
-            )
-
-            i = 1
-            for ev in button_pressed.ids:
-                cat_ob = Cat.fetch_cat(ev)
-                if cat_ob:
-                    # Shorten name if needed
-                    name = str(cat_ob.name)
-                    short_name = shorten_text_to_fit(name, 195, 26)
-
-                    self.cat_profile_buttons.append(
-                        IDImageButton(
-                            pygame.Rect(
-                                (
-                                    pressed_button_pos[0]
-                                    - (240 / 1600 * screen_x * i)
-                                    - 1,
-                                    pressed_button_pos[1] + 4,
-                                ),
-                                (232 / 1600 * screen_x, 60 / 1400 * screen_y),
-                            ),
-                            text=short_name,
-                            ids=ev,
-                            container=self.event_container,
-                            object_id="#events_cat_profile_button",
-                            layer_starting_height=2,
-                            manager=MANAGER,
-                        )
-                    )
-                    # There is only room for about four buttons.
-                    if i > 3:
-                        break
-                    i += 1
-
-        # If the button pressed does have its cat profile buttons open, just close the buttons.
-        else:
+        # if it does, clear the cat profile buttons
+        if self.open_involved_cat_button == button_pressed:
             self.open_involved_cat_button = None
             for ele in self.cat_profile_buttons:
-                ele.kill()
-            self.cat_profile_buttons = []
+                self.cat_profile_buttons[ele].kill()
+            self.cat_profile_buttons = {}
+            return
+
+        # If it doesn't have its buttons displayed, set the current open involved_cat_button to the pressed button,
+        # clear all other buttons, and open the cat profile buttons.
+        self.open_involved_cat_button = button_pressed
+        if self.involved_cat_container:
+            self.involved_cat_container.kill()
+
+        x_pos = 655
+        if game.settings["fullscreen"]:
+            y_pos = button_pressed.get_relative_rect()[1]
+        else:
+            y_pos = button_pressed.get_relative_rect()[1] * 2
+
+        self.involved_cat_container = UIModifiedScrollingContainer(
+            scale(pygame.Rect((20, y_pos), (890, 108))),
+            starting_height=3,
+            object_id="#involved_cat_container",
+            container=self.event_display,
+            manager=MANAGER,
+            allow_scroll_x=True
+        )
+
+        for i, cat_id in enumerate(button_pressed.ids):
+            cat_ob = Cat.fetch_cat(cat_id)
+            if cat_ob:
+                # Shorten name if needed
+                name = str(cat_ob.name)
+                short_name = shorten_text_to_fit(name, 195, 26)
+
+                self.cat_profile_buttons[f"profile_button{i}"] = IDImageButton(
+                    scale(pygame.Rect((x_pos, 4), (232, 60))),
+                    text=short_name,
+                    ids=cat_id,
+                    container=self.involved_cat_container,
+                    object_id="#events_cat_profile_button",
+                    layer_starting_height=1,
+                    manager=MANAGER,
+                )
+
+                x_pos += -255
+                if x_pos < 0:
+                    x_pos += 54
+
+        self.involved_cat_container.set_view_container_dimensions(
+            (self.involved_cat_container.get_relative_rect()[2], self.event_display.get_relative_rect()[3]))
+
+    def exit_screen(self):
+        self.event_display.kill()  # event display isn't put in the screen container due to lag issues
+        self.event_screen_container.kill()
 
     def update_display_events_lists(self):
         """
@@ -839,17 +346,185 @@ class EventsScreen(Screens):
         self.relation_events = [
             x for x in game.cur_events_list if "relation" in x.types
         ]
-        self.health_events = [x for x in game.cur_events_list if "health" in x.types]
+        self.health_events = [
+            x for x in game.cur_events_list if "health" in x.types
+        ]
         self.other_clans_events = [
             x for x in game.cur_events_list if "other_clans" in x.types
         ]
-        self.misc_events = [x for x in game.cur_events_list if "misc" in x.types]
+        self.misc_events = [
+            x for x in game.cur_events_list if "misc" in x.types
+        ]
 
-    def make_events_container(self):
-        """In its own function so that there is only one place the box size is set"""
-        self.event_container = UIModifiedScrollingContainer(
-            scale(pygame.Rect((432, 552), (1028, 700))),
-            object_id="#scroll_container",
-            starting_height=1,
-            manager=MANAGER,
-        )
+    def update_events_display(self):
+
+        # UPDATE CLAN INFO
+        self.clan_info["season"].set_text(f"Current season: {game.clan.current_season}")
+        if game.clan.age == 1:
+            self.clan_info["age"].set_text(f"Clan age: {game.clan.age} moon")
+        else:
+            self.clan_info["age"].set_text(f"Clan age: {game.clan.age} moons")
+
+        self.make_event_scrolling_container()
+
+        for ele in self.event_display_elements:
+            self.event_display_elements[ele].kill()
+        self.event_display_elements = {}
+
+        for ele in self.cat_profile_buttons:
+            self.cat_profile_buttons[ele].kill()
+        self.cat_profile_buttons = {}
+
+        for ele in self.involved_cat_buttons:
+            self.involved_cat_buttons[ele].kill()
+        self.involved_cat_buttons = {}
+
+        # Stop if Clan is new, so that events from previously loaded Clan don't show up
+        if game.clan.age == 0:
+            return
+
+        y_pos = 0
+
+        for i, event_object in enumerate(self.display_events):
+            # checking that text is a string
+            if not isinstance(event_object.text, str):
+                print(f"Incorrectly Formatted Event: {event_object.text}, {type(event_object)}")
+                continue
+
+            # TEXT BOX
+            self.event_display_elements[f"event{i}"] = pygame_gui.elements.UITextBox(
+                event_object.text,
+                scale(pygame.Rect((0, y_pos), (1018, -1))),
+                object_id=get_text_box_theme("#text_box_30_horizleft"),
+                starting_height=2,
+                container=self.event_display,
+                manager=MANAGER
+            )
+
+            if game.settings["fullscreen"]:
+                text_box_len = self.event_display_elements[f"event{i}"].get_relative_rect()[3]
+            else:
+                text_box_len = self.event_display_elements[f"event{i}"].get_relative_rect()[3] * 2
+
+            # SHADING
+            if i % 2 == 0:
+                image_path = "resources/images/shading"
+                if game.settings["dark mode"]:
+                    image_path += "_dark.png"
+                else:
+                    image_path += ".png"
+
+                if event_object.cats_involved:
+                    y_len = text_box_len + 125
+                else:
+                    y_len = text_box_len + 45
+
+                self.event_display_elements[f"shading{i}"] = pygame_gui.elements.UIImage(
+                    scale(pygame.Rect((0, y_pos), (1028, y_len))),
+                    image_cache.load_image(image_path),
+                    starting_height=1,
+                    object_id=f"shading{i}",
+                    container=self.event_display,
+                    manager=MANAGER
+                )
+                self.event_display_elements[f"shading{i}"].disable()
+
+            if event_object.cats_involved:
+                # INVOLVED CAT BUTTON
+                y_pos += text_box_len + 15
+
+                self.involved_cat_buttons[f"cat_button{i}"] = IDImageButton(
+                    scale(pygame.Rect((928, y_pos), (68, 68))),
+                    ids=event_object.cats_involved,
+                    layer_starting_height=3,
+                    object_id="#events_cat_button",
+                    container=self.event_display,
+                    manager=MANAGER
+                )
+
+                y_pos += 110
+            else:
+                y_pos += text_box_len + 45
+
+        # this HAS TO UPDATE before saved scroll position can be set
+        self.event_display.scrollable_container.update(1)
+
+        # don't ask me why we have to redefine these dimensions, we just do
+        # otherwise the scroll position save will break
+        self.event_display.set_dimensions(
+            (self.event_display.get_relative_rect()[2], self.event_display.get_relative_rect()[3]))
+
+        # set saved scroll position
+        if game.switches["saved_scroll_positions"].get(self.event_display_type):
+            self.event_display.vert_scroll_bar.set_scroll_from_start_percentage(
+                game.switches["saved_scroll_positions"][self.event_display_type]
+            )
+
+    def update_list_buttons(self):
+        """
+        re-enable all event tab buttons, then disable the currently selected tab
+        """
+        for ele in self.event_buttons:
+            self.event_buttons[ele].enable()
+
+        self.event_buttons[self.event_display_type].disable()
+
+    def on_use(self):
+        self.loading_screen_on_use(self.events_thread, self.timeskip_done)
+        pass
+
+    def timeskip_done(self):
+        """Various sorting and other tasks that must be done with the timeskip is over."""
+
+        game.switches["saved_scroll_positions"] = {}
+
+        if get_living_clan_cat_count(Cat) == 0:
+            GameOver("events screen")
+
+        self.update_display_events_lists()
+
+        self.event_display_type = "all"
+        self.event_buttons["all"].disable()
+
+        for tab in self.event_buttons:
+            if tab != "all":
+                self.event_buttons[tab].enable()
+
+        if not self.all_events:
+            self.all_events.append(
+                Single_Event("Nothing interesting happened this moon.")
+            )
+
+        self.display_events = self.all_events
+
+        if self.ceremony_events:
+            self.alert["ceremony"].show()
+        else:
+            self.alert["ceremony"].hide()
+
+        if self.birth_death_events:
+            self.alert["birth_death"].show()
+        else:
+            self.alert["birth_death"].hide()
+
+        if self.relation_events:
+            self.alert["relationship"].show()
+        else:
+            self.alert["relationship"].hide()
+
+        if self.health_events:
+            self.alert["health"].show()
+        else:
+            self.alert["health"].hide()
+
+        if self.other_clans_events:
+            self.alert["other_clans"].show()
+        else:
+            self.alert["other_clans"].hide()
+
+        if self.misc_events:
+            self.alert["misc"].show()
+        else:
+            self.alert["misc"].hide()
+
+        self.update_events_display()
